style: arrange directory structure
