+++ src/usecase/dump_global_variables.rs
