+++ src/usecase/mod.rs
@@ -1,0 +1,1 @@
+pub mod dump_global_variables;
